working on only 1 main.cpp for everyone

+++ src/NodeGraphControler.cpp
@@ -440,11 +440,11 @@ void NodeGraphControler::EditNodeParameters()
 
 void NodeGraphControler::HandlePinIO(size_t nodeIndex, size_t slotIndex, bool forOutput)
 {
-    if (IsIOUsed(nodeIndex, slotIndex, forOutput))
+    if (IsIOUsed(int(nodeIndex), int(slotIndex), forOutput))
 	{
             return;
 	}
-        ImGui::PushID(nodeIndex * 256 + slotIndex * 2 + (forOutput ? 1 : 0));
+        ImGui::PushID(int(nodeIndex * 256 + slotIndex * 2 + (forOutput ? 1 : 0)));
         bool pinned = IsIOPinned(nodeIndex, slotIndex, forOutput);
     ImGui::Checkbox("", &pinned);
         mEvaluationStages.SetIOPin(nodeIndex, slotIndex, forOutput, pinned);
@@ -480,11 +480,11 @@ void NodeGraphControler::NodeEdit()
                             {
                 continue;
 				}
-        ImGui::PushID(1717171 + nodeIndex);
+        ImGui::PushID(int(1717171 + nodeIndex));
         uint32_t parameterPair = (uint32_t(nodeIndex) << 16) + 0xDEAD;
         HandlePinIO(nodeIndex, 0, true);
         ImGui::SameLine();
-        Imogen::RenderPreviewNode(nodeIndex, *this);
+        Imogen::RenderPreviewNode(int(nodeIndex), *this);
         ImGui::PopID();
 		}
         PinnedEdit();
@@ -843,13 +843,13 @@ void NodeGraphControler::SetParameter(int nodeIndex,
         return;
     }
     size_t nodeType = mEvaluationStages.mStages[nodeIndex].mType;
-    int parameterIndex = GetParameterIndex(nodeType, parameterName.c_str());
+    int parameterIndex = GetParameterIndex(uint32_t(nodeType), parameterName.c_str());
     if (parameterIndex == -1)
     {
         return;
     }
-    ConTypes parameterType = GetParameterType(nodeType, parameterIndex);
-    size_t paramOffset = GetParameterOffset(nodeType, parameterIndex);
+    ConTypes parameterType = GetParameterType(uint32_t(nodeType), parameterIndex);
+    size_t paramOffset = GetParameterOffset(uint32_t(nodeType), parameterIndex);
     ParseStringToParameter(parameterValue, parameterType, &mEvaluationStages.mStages[nodeIndex].mParameters[paramOffset]);
     mEditingContext.SetTargetDirty(nodeIndex, Dirty::Parameter);
 }
